+++ src/mesh/cs_mesh_coarsen.cpp
@@ -2057,7 +2057,8 @@ cs_mesh_coarsen_simple(cs_mesh_t  *m,
 
   CS_FREE(vtx_o2n);
 
-  m->n_b_faces_all = m->n_b_faces;
+  m->n_g_b_faces_appended = 0;
+  m->n_b_faces_appended = 0;
   m->modified |= (CS_MESH_MODIFIED | CS_MESH_MODIFIED_BALANCE);
 
   /* Rebuild auxiliary information */